Simplify the domToModel call in `paste.ts` (#2040)
* add more changes

* fix build

* fix test
