+++ packages-content-model/roosterjs-content-model-editor/lib/publicApi/utils/paste.ts
@@ -4,11 +4,7 @@ import { FormatWithContentModelContext } from '../../publicTypes/parameter/Forma
 import { IContentModelEditor } from '../../publicTypes/IContentModelEditor';
 import { mergeModel } from '../../modelApi/common/mergeModel';
 import { NodePosition } from 'roosterjs-editor-types';
-import {
-    ContentModelBlockFormat,
-    ContentModelDocument,
-    FormatParser,
-} from 'roosterjs-content-model-types';
+import { ContentModelDocument } from 'roosterjs-content-model-types';
 import ContentModelBeforePasteEvent, {
     ContentModelBeforePasteEventData,
 } from '../../publicTypes/event/ContentModelBeforePasteEvent';
@@ -70,20 +66,7 @@ export default function paste(
         eventData
     );
 
-    const pasteModel = domToContentModel(fragment, {
-        ...domToModelOption,
-        additionalFormatParsers: {
-            ...domToModelOption.additionalFormatParsers,
-            block: [
-                ...(domToModelOption.additionalFormatParsers?.block || []),
-                ...(applyCurrentFormat ? [blockElementParser] : []),
-            ],
-            listLevel: [
-                ...(domToModelOption.additionalFormatParsers?.listLevel || []),
-                ...(applyCurrentFormat ? [blockElementParser] : []),
-            ],
-        },
-    });
+    const pasteModel = domToContentModel(fragment, domToModelOption);
 
     if (pasteModel) {
         formatWithContentModel(
@@ -214,17 +197,3 @@ function triggerPluginEventAndCreatePasteFragment(
 
     return pluginEvent;
 }
-
-/**
- * For block elements that have background color style, remove the background color when user selects the merge current format
- * paste option
- */
-const blockElementParser: FormatParser<ContentModelBlockFormat> = (
-    format: ContentModelBlockFormat,
-    element: HTMLElement
-) => {
-    if (element.style.backgroundColor) {
-        element.style.backgroundColor = '';
-        delete format.backgroundColor;
-    }
-};
